requirements.txt
attrs==19.3.0
beautifulsoup4==4.9.0
bokeh==2.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
geopandas==0.7.0
google-auth==1.13.1
gspread==3.4.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
Pillow==7.1.1
plotly==4.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyproj==2.6.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
rsa==4.0
Shapely==1.7.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.8
Werkzeug==1.0.1
